refactor: a-train branding
